+++ src/forze_awskms/_logging.py
@@ -1,0 +1,18 @@
+"""Constants for the forze_awskms package."""
+
+from enum import StrEnum
+from typing import Final, final
+
+# ----------------------- #
+
+
+@final
+class ForzeAwsKmsLogger(StrEnum):
+    """Forze AWS KMS logger names."""
+
+    KERNEL = "forze_awskms.kernel"
+
+
+# ....................... #
+
+FORZE_AWSKMS_LOGGER_NAMES: Final = list(map(str, ForzeAwsKmsLogger))